Bluelink: make language configurable
